feat(landing): activate retail IA A/B cohort (landing_retail_ia_2026)
Deploy 50/50 middleware cookie split on / with retail educational variant vs
control sovereign IA. Adds Firestore cohort telemetry, admin dashboard pane,
and compliance-safe copy. Canonical SEO unchanged.
